# Simplify renderError