Rather than an additional special case of `/_error` in `next/client` to handle calling `getInitialProps`, lets just do it all in the router. We do this by treating `renderError` as a special navigation.

It might make sense to squelch the navigation events as a side effect, but that's left unaddressed as of now.