@ijjk 2020-05-20
# Stats from current PR

<details>
<summary><strong>Default Server Mode</strong> (Increase detected ⚠️)</summary>

<br/>

<details>
<summary><strong>General</strong> Overall decrease ✓</summary>

|  | zeit/next.js canary  | azukaru/next.js simplify-render-error | Change |
| - | - | - | - |
| buildDuration | 10.8s | 11s | ⚠️ +180ms |
| nodeModulesSize | 61.8 MB | 61.8 MB | -1.21 kB |

</details>

<details>
<summary><strong>Page Load Tests</strong> Overall decrease ⚠️</summary>

|  | zeit/next.js canary  | azukaru/next.js simplify-render-error | Change |
| - | - | - | - |
| / failed reqs | 0 | 0 | ✓ |
| / total time (seconds) | 1.812 | 1.878 | ⚠️ +0.07 |
| / avg req/sec | 1379.87 | 1330.9 | -48.97 |
| /error-in-render failed reqs | 0 | 0 | ✓ |
| /error-in-render total time (seconds) | 1.2 | 1.192 | -0.01 |
| /error-in-render avg req/sec | 2083.49 | 2097.77 | ⚠️ +14.28 |

</details>

<details>
<summary><strong>Client Bundles (main, webpack, commons)</strong> Overall increase ⚠️</summary>

|  | zeit/next.js canary  | azukaru/next.js simplify-render-error | Change |
| - | - | - | - |
| main-HASH.js gzip | 6.49 kB | 6.41 kB | -88 B |
| webpack-HASH.js gzip | 746 B | 746 B | ✓ |
| de003c3a9d30..fef6.js gzip | 10.5 kB | N/A | N/A |
| framework.HASH.js gzip | 39.1 kB | 39.1 kB | ✓ |
| de003c3a9d30..7879.js gzip | N/A | 10.6 kB | N/A |
| Overall change | 56.9 kB | 56.9 kB | ⚠️ +45 B |

</details>

<details>
<summary><strong>Client Bundles (main, webpack, commons) Modern</strong> Overall decrease ✓</summary>

|  | zeit/next.js canary  | azukaru/next.js simplify-render-error | Change |
| - | - | - | - |
| main-HASH.module.js gzip | 5.58 kB | 5.51 kB | -71 B |
| webpack-HASH..dule.js gzip | 746 B | 746 B | ✓ |
| de003c3a9d30..dule.js gzip | 6.92 kB | 6.94 kB | ⚠️ +22 B |
| framework.HA..dule.js gzip | 39.1 kB | 39.1 kB | ✓ |
| Overall change | 52.4 kB | 52.3 kB | -49 B |

</details>

<details>
<summary><strong>Legacy Client Bundles (polyfills)</strong></summary>

|  | zeit/next.js canary  | azukaru/next.js simplify-render-error | Change |
| - | - | - | - |
| polyfills-HASH.js gzip | 26.3 kB | 26.3 kB | ✓ |
| Overall change | 26.3 kB | 26.3 kB | ✓ |

</details>

<details>
<summary><strong>Client Pages</strong></summary>

|  | zeit/next.js canary  | azukaru/next.js simplify-render-error | Change |
| - | - | - | - |
| _app.js gzip | 1.26 kB | 1.26 kB | ✓ |
| _error.js gzip | 3.25 kB | 3.25 kB | ✓ |
| hooks.js gzip | 881 B | 881 B | ✓ |
| index.js gzip | 222 B | 222 B | ✓ |
| link.js gzip | 2.06 kB | 2.06 kB | ✓ |
| routerDirect.js gzip | 279 B | 279 B | ✓ |
| withRouter.js gzip | 278 B | 278 B | ✓ |
| Overall change | 8.22 kB | 8.22 kB | ✓ |

</details>

<details>
<summary><strong>Client Pages Modern</strong></summary>

|  | zeit/next.js canary  | azukaru/next.js simplify-render-error | Change |
| - | - | - | - |
| _app.module.js gzip | 604 B | 604 B | ✓ |
| _error.module.js gzip | 2.04 kB | 2.04 kB | ✓ |
| hooks.module.js gzip | 383 B | 383 B | ✓ |
| index.module.js gzip | 223 B | 223 B | ✓ |
| link.module.js gzip | 1.52 kB | 1.52 kB | ✓ |
| routerDirect..dule.js gzip | 281 B | 281 B | ✓ |
| withRouter.m..dule.js gzip | 278 B | 278 B | ✓ |
| Overall change | 5.33 kB | 5.33 kB | ✓ |

</details>

<details>
<summary><strong>Client Build Manifests</strong></summary>

|  | zeit/next.js canary  | azukaru/next.js simplify-render-error | Change |
| - | - | - | - |
| _buildManifest.js gzip | 61 B | 61 B | ✓ |
| _buildManife..dule.js gzip | 61 B | 61 B | ✓ |
| Overall change | 122 B | 122 B | ✓ |

</details>

<details>
<summary><strong>Rendered Page Sizes</strong> Overall increase ⚠️</summary>

|  | zeit/next.js canary  | azukaru/next.js simplify-render-error | Change |
| - | - | - | - |
| index.html gzip | 928 B | 927 B | -1 B |
| link.html gzip | 937 B | 938 B | ⚠️ +1 B |
| withRouter.html gzip | 924 B | 925 B | ⚠️ +1 B |
| Overall change | 2.79 kB | 2.79 kB | ⚠️ +1 B |

</details>

#### Diffs
<details>
<summary>Diff for <strong>de003c3a9d30..95.module.js</strong></summary>

```diff
@@ -676,6 +676,34 @@ class Router {
     return this.change('replaceState', url, as, options);
   }
 
+  async _handleError(err, pathname, query, props) {
+    var {
+      page: Component
+    } = await this.fetchComponent('/_error');
+    var routeInfo = {
+      Component,
+      err,
+      error: err,
+      props
+    };
+
+    if (!props) {
+      try {
+        props = await this.getInitialProps(Component, {
+          err,
+          pathname,
+          query
+        });
+        routeInfo.props = props;
+      } catch (gipErr) {
+        console.error('Error in error page `getInitialProps`: ', gipErr);
+        routeInfo.props = {};
+      }
+    }
+
+    return routeInfo;
+  }
+
   change(method, _url, _as, options) {
     return new Promise((resolve, reject) => {
       if (!options._h) {
@@ -759,9 +787,12 @@ class Router {
         }
       }
 
-      Router.events.emit('routeChangeStart', as); // If shallow is true and the route exists in the router cache we reuse the previous result
+      Router.events.emit('routeChangeStart', as); // WARNING: `_e` is an internal option.  Your app should _never_ use
+      // this property. It may change at any time without notice.
 
-      this.getRouteInfo(route, pathname, query, as, shallow).then(routeInfo => {
+      var routeInfoPromise = options._e ? this._handleError(options._e.err, pathname, query, options._e.props) : // If shallow is true and the route exists in the router cache we reuse the previous result
+      this.getRouteInfo(route, pathname, query, as, shallow);
+      routeInfoPromise.then(routeInfo => {
         var {
           error
         } = routeInfo;
@@ -839,31 +870,7 @@ class Router {
           });
         }
 
-        resolve(this.fetchComponent('/_error').then(res => {
-          var {
-            page: Component
-          } = res;
-          var routeInfo = {
-            Component,
-            err
-          };
-          return new Promise(resolve => {
-            this.getInitialProps(Component, {
-              err,
-              pathname,
-              query
-            }).then(props => {
-              routeInfo.props = props;
-              routeInfo.error = err;
-              resolve(routeInfo);
-            }, gipErr => {
-              console.error('Error in error page `getInitialProps`: ', gipErr);
-              routeInfo.error = err;
-              routeInfo.props = {};
-              resolve(routeInfo);
-            });
-          });
-        }).catch(err => handleError(err, true)));
+        resolve(this._handleError(err, pathname, query).catch(err => handleError(err, true)));
       });
     };
```
</details>
<details>
<summary>Diff for <strong>de003c3a9d30..af23efe53.js</strong></summary>

```diff
@@ -479,12 +479,12 @@ function mitt() {
 "use strict";
 
 
+var _slicedToArray = __webpack_require__("J4zp");
+
 var _regeneratorRuntime = __webpack_require__("o0o1");
 
 var _asyncToGenerator = __webpack_require__("yXPU");
 
-var _slicedToArray = __webpack_require__("J4zp");
-
 var _classCallCheck = __webpack_require__("lwsE");
 
 var _createClass = __webpack_require__("W8MJ");
@@ -801,6 +801,71 @@ var Router = /*#__PURE__*/function () {
       var options = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : {};
       return this.change('replaceState', url, as, options);
     }
+  }, {
+    key: "_handleError",
+    value: function () {
+      var _handleError2 = _asyncToGenerator( /*#__PURE__*/_regeneratorRuntime.mark(function _callee(err, pathname, query, props) {
+        var _yield$this$fetchComp, Component, routeInfo;
+
+        return _regeneratorRuntime.wrap(function _callee$(_context) {
+          while (1) {
+            switch (_context.prev = _context.next) {
+              case 0:
+                _context.next = 2;
+                return this.fetchComponent('/_error');
+
+              case 2:
+                _yield$this$fetchComp = _context.sent;
+                Component = _yield$this$fetchComp.page;
+                routeInfo = {
+                  Component: Component,
+                  err: err,
+                  error: err,
+                  props: props
+                };
+
+                if (props) {
+                  _context.next = 17;
+                  break;
+                }
+
+                _context.prev = 6;
+                _context.next = 9;
+                return this.getInitialProps(Component, {
+                  err: err,
+                  pathname: pathname,
+                  query: query
+                });
+
+              case 9:
+                props = _context.sent;
+                routeInfo.props = props;
+                _context.next = 17;
+                break;
+
+              case 13:
+                _context.prev = 13;
+                _context.t0 = _context["catch"](6);
+                console.error('Error in error page `getInitialProps`: ', _context.t0);
+                routeInfo.props = {};
+
+              case 17:
+                return _context.abrupt("return", routeInfo);
+
+              case 18:
+              case "end":
+                return _context.stop();
+            }
+          }
+        }, _callee, this, [[6, 13]]);
+      }));
+
+      function _handleError(_x, _x2, _x3, _x4) {
+        return _handleError2.apply(this, arguments);
+      }
+
+      return _handleError;
+    }()
   }, {
     key: "change",
     value: function change(method, _url, _as, options) {
@@ -892,9 +957,12 @@ var Router = /*#__PURE__*/function () {
           }
         }
 
-        Router.events.emit('routeChangeStart', as); // If shallow is true and the route exists in the router cache we reuse the previous result
+        Router.events.emit('routeChangeStart', as); // WARNING: `_e` is an internal option.  Your app should _never_ use
+        // this property. It may change at any time without notice.
 
-        _this2.getRouteInfo(route, pathname, query, as, shallow).then(function (routeInfo) {
+        var routeInfoPromise = options._e ? _this2._handleError(options._e.err, pathname, query, options._e.props) : // If shallow is true and the route exists in the router cache we reuse the previous result
+        _this2.getRouteInfo(route, pathname, query, as, shallow);
+        routeInfoPromise.then(function (routeInfo) {
           var error = routeInfo.error;
 
           if (error && error.cancelled) {
@@ -975,29 +1043,7 @@ var Router = /*#__PURE__*/function () {
             });
           }
 
-          resolve(_this3.fetchComponent('/_error').then(function (res) {
-            var Component = res.page;
-            var routeInfo = {
-              Component: Component,
-              err: err
-            };
-            return new Promise(function (resolve) {
-              _this3.getInitialProps(Component, {
-                err: err,
-                pathname: pathname,
-                query: query
-              }).then(function (props) {
-                routeInfo.props = props;
-                routeInfo.error = err;
-                resolve(routeInfo);
-              }, function (gipErr) {
-                console.error('Error in error page `getInitialProps`: ', gipErr);
-                routeInfo.error = err;
-                routeInfo.props = {};
-                resolve(routeInfo);
-              });
-            });
-          })["catch"](function (err) {
+          resolve(_this3._handleError(err, pathname, query)["catch"](function (err) {
             return handleError(err, true);
           }));
         });
@@ -1158,11 +1204,11 @@ var Router = /*#__PURE__*/function () {
   }, {
     key: "fetchComponent",
     value: function () {
-      var _fetchComponent = _asyncToGenerator( /*#__PURE__*/_regeneratorRuntime.mark(function _callee(route) {
+      var _fetchComponent = _asyncToGenerator( /*#__PURE__*/_regeneratorRuntime.mark(function _callee2(route) {
         var cancelled, cancel, componentResult, error;
-        return _regeneratorRuntime.wrap(function _callee$(_context) {
+        return _regeneratorRuntime.wrap(function _callee2$(_context2) {
           while (1) {
-            switch (_context.prev = _context.next) {
+            switch (_context2.prev = _context2.next) {
               case 0:
                 cancelled = false;
 
@@ -1171,14 +1217,14 @@ var Router = /*#__PURE__*/function () {
                 };
 
                 route = delBasePath(route);
-                _context.next = 5;
+                _context2.next = 5;
                 return this.pageLoader.loadPage(route);
 
               case 5:
-                componentResult = _context.sent;
+                componentResult = _context2.sent;
 
                 if (!cancelled) {
-                  _context.next = 10;
+                  _context2.next = 10;
                   break;
                 }
 
@@ -1191,17 +1237,17 @@ var Router = /*#__PURE__*/function () {
                   this.clc = null;
                 }
 
-                return _context.abrupt("return", componentResult);
+                return _context2.abrupt("return", componentResult);
 
               case 12:
               case "end":
-                return _context.stop();
+                return _context2.stop();
             }
           }
-        }, _callee, this);
+        }, _callee2, this);
       }));
 
-      function fetchComponent(_x) {
+      function fetchComponent(_x5) {
         return _fetchComponent.apply(this, arguments);
       }
```
</details>
<details>
<summary>Diff for <strong>main-HASH.js</strong></summary>

```diff
@@ -606,8 +606,7 @@ function _render() {
 }
 
 function renderError(props) {
-  var App = props.App,
-      err = props.err; // In development runtime errors are caught by our overlay
+  var err = props.err; // In development runtime errors are caught by our overlay
   // In production we catch runtime errors using componentDidCatch which will trigger renderError
 
   if (false) {}
@@ -616,31 +615,11 @@ function renderError(props) {
 
 
   console.error(err);
-  return pageLoader.loadPage('/_error').then(function (_ref7) {
-    var ErrorComponent = _ref7.page;
-    // In production we do a normal render with the `ErrorComponent` as component.
-    // If we've gotten here upon initial render, we can use the props from the server.
-    // Otherwise, we need to call `getInitialProps` on `App` before mounting.
-    var AppTree = wrapApp(App);
-    var appCtx = {
-      Component: ErrorComponent,
-      AppTree: AppTree,
-      router: router,
-      ctx: {
-        err: err,
-        pathname: page,
-        query: query,
-        asPath: asPath,
-        AppTree: AppTree
-      }
-    };
-    return Promise.resolve(props.props ? props.props : (0, _utils.loadGetInitialProps)(App, appCtx)).then(function (initProps) {
-      return doRender((0, _extends2["default"])((0, _extends2["default"])({}, props), {}, {
-        err: err,
-        Component: ErrorComponent,
-        props: initProps
-      }));
-    });
+  return router.change('replaceState', page, asPath, {
+    _e: {
+      err: err,
+      props: props.props
+    }
   });
 } // If hydrate does not exist, eg in preact.
 
@@ -715,8 +694,8 @@ function clearMarks() {
   });
 }
 
-function AppContainer(_ref8) {
-  var children = _ref8.children;
+function AppContainer(_ref7) {
+  var children = _ref7.children;
   return /*#__PURE__*/_react["default"].createElement(Container, {
     fn: function fn(error) {
       return renderError({
@@ -749,13 +728,13 @@ function doRender(_x2) {
 }
 
 function _doRender() {
-  _doRender = _asyncToGenerator( /*#__PURE__*/_regeneratorRuntime.mark(function _callee3(_ref9) {
+  _doRender = _asyncToGenerator( /*#__PURE__*/_regeneratorRuntime.mark(function _callee3(_ref8) {
     var App, Component, props, err, appProps, elem;
     return _regeneratorRuntime.wrap(function _callee3$(_context3) {
       while (1) {
         switch (_context3.prev = _context3.next) {
           case 0:
-            App = _ref9.App, Component = _ref9.Component, props = _ref9.props, err = _ref9.err;
+            App = _ref8.App, Component = _ref8.Component, props = _ref8.props, err = _ref8.err;
             Component = Component || lastAppProps.Component;
             props = props || lastAppProps.props;
             appProps = (0, _extends2["default"])((0, _extends2["default"])({}, props), {}, {
```
</details>
<details>
<summary>Diff for <strong>main-HASH.module.js</strong></summary>

```diff
@@ -447,7 +447,6 @@ async function render(props) {
 
 function renderError(props) {
   var {
-    App,
     err
   } = props; // In development runtime errors are caught by our overlay
   // In production we catch runtime errors using componentDidCatch which will trigger renderError
@@ -458,31 +457,11 @@ function renderError(props) {
 
 
   console.error(err);
-  return pageLoader.loadPage('/_error').then((_ref4) => {
-    var {
-      page: ErrorComponent
-    } = _ref4;
-    // In production we do a normal render with the `ErrorComponent` as component.
-    // If we've gotten here upon initial render, we can use the props from the server.
-    // Otherwise, we need to call `getInitialProps` on `App` before mounting.
-    var AppTree = wrapApp(App);
-    var appCtx = {
-      Component: ErrorComponent,
-      AppTree,
-      router,
-      ctx: {
-        err,
-        pathname: page,
-        query,
-        asPath,
-        AppTree
-      }
-    };
-    return Promise.resolve(props.props ? props.props : (0, _utils.loadGetInitialProps)(App, appCtx)).then(initProps => doRender((0, _extends2.default)((0, _extends2.default)({}, props), {}, {
-      err,
-      Component: ErrorComponent,
-      props: initProps
-    })));
+  return router.change('replaceState', page, asPath, {
+    _e: {
+      err: err,
+      props: props.props
+    }
   });
 } // If hydrate does not exist, eg in preact.
 
@@ -553,10 +532,10 @@ function clearMarks() {
   ['beforeRender', 'afterHydrate', 'afterRender', 'routeChange'].forEach(mark => performance.clearMarks(mark));
 }
 
-function AppContainer(_ref5) {
+function AppContainer(_ref4) {
   var {
     children
-  } = _ref5;
+  } = _ref4;
   return /*#__PURE__*/_react.default.createElement(Container, {
     fn: error => renderError({
       App,
@@ -578,13 +557,13 @@ var wrapApp = App => props => {
   return /*#__PURE__*/_react.default.createElement(AppContainer, null, /*#__PURE__*/_react.default.createElement(App, appProps));
 };
 
-async function doRender(_ref6) {
+async function doRender(_ref5) {
   var {
     App,
     Component,
     props,
     err
-  } = _ref6;
+  } = _ref5;
   Component = Component || lastAppProps.Component;
   props = props || lastAppProps.props;
   var appProps = (0, _extends2.default)((0, _extends2.default)({}, props), {}, {
```
</details>
<details>
<summary>Diff for <strong>index.html</strong></summary>

```diff
@@ -1 +1 @@
…